+++ lib/types/primitives/team.dart
@@ -4,3 +4,10 @@ class Team {
   final String displayName;
   Team(this.name, this.id, this.displayName);
 }
+
+class ModuleStatistics {
+  double deviation;
+  int total;
+  int normal;
+  ModuleStatistics(this.deviation, this.total, this.normal);
+}